requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
absl-py=0.10.0=py38h32f6830_1
aiohttp=3.7.2=py38h4d0b108_0
appnope=0.1.0=py38_1001
argon2-cffi=20.1.0=py38haf1e3a3_1
astroid=2.4.2=py38_0
async-timeout=3.0.1=py_1000
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.1=py38hecd8cb5_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py_1
brotlipy=0.7.0=py38h9ed2024_1003
c-ares=1.16.1=haf1e3a3_3
ca-certificates=2021.1.19=hecd8cb5_1
cachetools=4.1.1=py_0
certifi=2020.12.5=py38hecd8cb5_0
cffi=1.14.3=py38h2125817_2
chardet=3.0.4=py38hecd8cb5_1003
click=7.1.2=py_0
colorama=0.4.3=py_0
conda=4.9.2=py38hecd8cb5_0
conda-package-handling=1.7.2=py38h22f3db7_0
cryptography=3.2.1=py38hbcfaee0_1
cycler=0.10.0=py38_0
dbus=1.13.18=h18a8e69_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docopt=0.6.2=pypi_0
entrypoints=0.3=py38_0
et_xmlfile=1.0.1=py_1001
expat=2.2.10=hb1e8313_2
filelock=3.0.12=pypi_0
flask=1.1.2=pypi_0
freetype=2.10.4=ha233b18_0
fsspec=0.8.4=py_0
future=0.18.2=py38h50d1736_2
gettext=0.19.8.1=hb0f4f8b_2
git=2.23.0=pl526h6951d83_0
glib=2.66.1=h9bbe63b_0
google-auth=1.23.0=pyhd8ed1ab_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.31.0=py38h7580e61_0
h5py=2.10.0=py38h0601b69_1
hdf5=1.10.6=hdbbcd12_0
icu=58.2=h0a44026_3
idna=2.10=py_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2019.4=233
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38h01d92e1_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
isort=5.6.4=py_0
itsdangerous=1.1.0=pypi_0
jdcal=1.4.1=py_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.0=py38h23ab428_0
krb5=1.18.2=h75d18d8_0
langdetect=1.0.8=pypi_0
lazy-object-proxy=1.4.3=py38h1de35cc_0
lcms2=2.11=h92f6f08_0
libcurl=7.71.1=h8a08a2b_1
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libmklml=2019.0.5=0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.13.0.1=hab81aa3_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.9.0=ha12b0ac_1
libtiff=4.1.0=hcb84e12_1
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h79c402e_3
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py38h1de35cc_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h181983e_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h1de35cc_1001
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
multidict=4.7.5=py38h4d0b108_2
natsort=7.1.1=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.2=pyhd3eb1b0_0
networkx=2.5=pypi_0
ninja=1.10.1=py38h879752b_0
nltk=3.5=py_0
nose=1.3.7=py38_1004
notebook=6.1.4=py38_0
num2words=0.5.10=pypi_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1k=h9ed2024_0
packaging=20.4=py_0
pandas=1.1.3=py38hb1e8313_0
pandoc=2.11=h0dc7051_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.7.0=py_0
pcre=8.44=hb1e8313_0
perl=5.26.2=h4e221da_0
pexpect=4.8.0=py38_1
pickleshare=0.7.5=py38_1001
pillow=8.0.1=py38h5270095_0
pip=20.2.4=py38hecd8cb5_0
poutyne=1.1=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.13.0.1=py38hb1e8313_1
ptyprocess=0.6.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycosat=0.6.3=py38h1de35cc_1
pycparser=2.20=py_2
pygments=2.7.2=pyhd3eb1b0_0
pyjwt=1.7.1=py_0
pylint=2.6.0=py38_0
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.17.3=py38haf1e3a3_0
pyscm-ml=1.0.3=pypi_0
pysocks=1.7.1=py38_1
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python.app=2=py38_10
python_abi=3.8=1_cp38
pytorch=1.6.0=cpu_py38hd70000b_0
pytorch-lightning=1.0.6=pyhd8ed1ab_0
pytz=2020.1=py_0
pyyaml=5.3.1=py38h94c058a_1
pyzmq=19.0.2=py38hb1e8313_1
qt=5.9.7=h468cd18_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h1de35cc_0
regex=2020.10.15=py38haf1e3a3_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rope=0.18.0=py_0
rsa=4.6=pyh9f0ad1d_0
ruamel_yaml=0.15.87=py38haf1e3a3_1
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=py38h959d312_0
scipy=1.5.2=py38h2515648_0
seaborn=0.11.0=py_0
selenium=3.141.0=pypi_0
send2trash=1.5.0=py38_0
setuptools=50.3.1=py38hecd8cb5_1
sip=4.19.8=py38h0a44026_0
six=1.15.0=py38hecd8cb5_0
soupsieve=2.0.1=py_0
sqlite=3.33.0=hffcf06c_0
tensorboard=2.4.0=pyhd8ed1ab_0
tensorboard-plugin-wit=1.7.0=pyh9f0ad1d_0
terminado=0.9.1=py38_0
test-tube=0.7.5=pypi_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
tokenizers=0.10.1=pypi_0
toml=0.10.1=py_0
torchvision=0.1.8=pypi_0
tornado=6.0.4=py38h1de35cc_1
tqdm=4.51.0=pyhd3eb1b0_0
traitlets=5.0.5=py_0
transformers=4.4.2=pypi_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
unidecode=1.1.1=py_0
urllib3=1.25.11=py_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pyh9f0ad1d_0
wget=1.20.1=h051b688_0
wheel=0.35.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
wrapt=1.11.2=py38h1de35cc_0
xlrd=1.2.0=py_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
yarl=1.6.3=py38h7e3306e_0
zeromq=4.3.3=hb1e8313_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
